+++ integrations/github/src/actions/index.ts
@@ -1,5 +1,5 @@
 import { findTarget, findIssuesOrPullRequests } from './find-target'
-import { sendRawGraphqlQuery } from './send-custom-query'
+import { sendRawGraphqlQuery } from './send-raw-graphql-query'
 
 export default {
   findTarget,
